RavenDB-12376: Refactor and add errno string handling in ThrowLastError (still linux x64 version only)
